pyproject.toml
[tool.poetry]
name = "wardrobe"
version = "0.1.0"
description = ""
authors = ["Pyunghyuk Yoo <yoophi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "1.1.2"
authlib = { git = "https://github.com/lepture/authlib.git", branch = "master" }
flask-sqlalchemy = "^2.5.1"
loguru = "^0.5.3"
flask_migrate = "^3.0.0"
email_validator = "^1.1.2"
flask_cors = "^3.0.10"
flask_login = "^0.5.0"
flask_marshmallow = "^0.14.0"
flask_user = "^1.0.2"
flask_swagger = "^0.2.14"
requests = "^2.25.1"
psycopg2-binary = "^2.8.6"
flask-social-login = "0.2.2"
flask-assets = "2.0"
cryptography = "3.3.2"
gunicorn = "^20.1.0"
ca-util = "^0.1.0"


[tool.poetry.dev-dependencies]
twine = "^3.4.1"
bump2version = "^1.0.1"
wheel = "^0.36.2"
watchdog = "^2.1.2"
flake8 = "^3.9.2"
black = "^21.6b0"
coverage = "^5.5"
Sphinx = "^4.0.2"
pytest = "^6.2.4"
pytest-runner = "^5.3.1"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
